Move displaying lambda logs to logs module
