+++ chalice/logs.py
@@ -4,8 +4,22 @@ This module provides APIs for searching, interacting
 with the logs generated by AWS Lambda.
 
 """
-from typing import Any, Optional, Iterator, Dict  # noqa
 import datetime
+
+from typing import Any, Optional, Iterator, Dict  # noqa
+from botocore.session import Session  # noqa
+
+
+def show_lambda_logs(session, lambda_arn, max_entries,
+                     include_lambda_messages):
+    # type: (Session, str, int, bool) -> None
+    client = session.create_client('logs')
+    retriever = LogRetriever.create_from_arn(client, lambda_arn)
+    events = retriever.retrieve_logs(
+        include_lambda_messages=include_lambda_messages,
+        max_entries=max_entries)
+    for event in events:
+        print event['timestamp'], event['logShortId'], event['message'].strip()
 
 
 class LogRetriever(object):
